# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Azile AI MDM API
  version: "3.1.0"
  description: |
    Azile AI MDM is a hub-and-spoke Master Data Management platform built on Rust microservices.

    ## Services
    | Service | Port | Purpose |
    |---------|------|---------|
    | API Gateway | 8080 | Single entry point; auth + rate-limit + proxy |
    | MDM Core | 8081 | Entities, matching, merge, approval, admin config |
    | AI Service | 8082 | LLM copilot, embeddings, semantic matching, RAG |
    | Policy Service | 8083 | Policy evaluation, GDPR, consent |
    | Ingest Service | 8084 | Batch/CSV/REST ingestion, source systems |
    | Search Service | 8085 | Full-text search, autocomplete |
    | Notification Service | 8086 | In-app notifications, webhooks, email |

    ## Authentication
    All protected endpoints require a JWT Bearer token obtained from `POST /auth/login`.
    SCIM 2.0 endpoints use a separate SCIM bearer token obtained from `POST /scim/tokens`.

    Standard request headers:
    - `Authorization: Bearer <jwt>` â€” JWT from login
    - `x-tenant-id` â€” Tenant UUID (injected by gateway; required on direct calls)
    - `x-correlation-id` â€” Optional idempotency / tracing ID

  contact:
    name: Azile MDM Support
    email: ajit.more@lithytree.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: API Gateway (recommended entry point)
  - url: http://localhost:8081
    description: MDM Core (direct)
...
security:
  - bearerAuth: []
...
paths:
...
  /ingest/batch:
    post:
      tags: [Ingestion]
      summary: Ingest a batch of entity records (JSON)
      servers:
        - url: http://localhost:8084
          description: Ingest Service
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IngestBatchRequest" }
      responses:
        "202":
          description: Batch accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  job_id: { type: string, format: uuid }
  /ingest/csv:
    post:
      tags: [Ingestion]
      summary: Ingest entities from a CSV file upload
      servers:
        - url: http://localhost:8084
          description: Ingest Service
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, entity_type]
              properties:
                file:
                  type: string
                  format: binary
                entity_type: { type: string }
                source_system: { type: string }
                mapping:
                  type: object
                  additionalProperties: { type: string }
                  description: "CSV column â†’ attribute code mapping"
      responses:
        "202":
          description: CSV accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  job_id: { type: string, format: uuid }
  /ingest/jobs:
    get:
      tags: [Ingestion]
      summary: List ingest jobs
      servers:
        - url: http://localhost:8084
          description: Ingest Service
      parameters:
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PageSize" }
      responses:
        "200":
          description: Jobs list
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
  /ingest/jobs/{id}:
    get:
      tags: [Ingestion]
      summary: Get ingest job status and error report
      servers:
        - url: http://localhost:8084
          description: Ingest Service
      parameters:
        - { $ref: "#/components/parameters/Id" }
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
  /source-systems:
    get:
      tags: [Source Systems]
      summary: List source system configurations
      servers:
        - url: http://localhost:8084
          description: Ingest Service
      responses:
        "200":
          description: Source systems list
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }

    post:
      tags: [Source Systems]
      summary: Create a source system
      servers:
        - url: http://localhost:8084
          description: Ingest Service
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSourceSystemRequest" }
      responses:
        "201":
          description: Source system created
  /source-systems/{id}:
    put:
      tags: [Source Systems]
      summary: Update a source system
      servers:
        - url: http://localhost:8084
          description: Ingest Service
      parameters:
        - { $ref: "#/components/parameters/Id" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSourceSystemRequest" }
      responses:
        "200":
          description: Source system updated

    delete:
      tags: [Source Systems]
      summary: Delete a source system
      servers:
        - url: http://localhost:8084
          description: Ingest Service
      parameters:
        - { $ref: "#/components/parameters/Id" }
      responses:
        "204":
          description: Deleted
  /source-systems/{id}/test:
    post:
      tags: [Source Systems]
      summary: Test connectivity to a source system
      servers:
        - url: http://localhost:8084
          description: Ingest Service
      parameters:
        - { $ref: "#/components/parameters/Id" }
      responses:
        "200":
          description: Connection test result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  latency_ms: { type: integer }
                  error: { type: string, nullable: true }

# â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
# COMPONENTS
# â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT obtained from POST /auth/login

    scimBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: SCIM Token
      description: SCIM bearer token obtained from POST /scim/tokens
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Page:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, default: 20, minimum: 1, maximum: 200 }
...
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success: { type: boolean }
        data: {}
        message: { type: string }
        error: { type: string, nullable: true }
...
    IngestBatchRequest:
      type: object
      required: [entity_type, records]
      properties:
        entity_type: { type: string }
        source_system: { type: string }
        records:
          type: array
          items:
            type: object
            properties:
              external_id: { type: string }
              attributes: { type: object, additionalProperties: true }
        on_conflict: { type: string, enum: [update, skip, error], default: update }
    CreateSourceSystemRequest:
      type: object
      required: [name, code, system_type]
      properties:
        name: { type: string }
        code: { type: string }
        system_type:
          type: string
          enum: [rest_api, database, kafka, file_upload, salesforce, sap, custom]
        connection_config:
          type: object
          additionalProperties: true
          description: System-type-specific config (URLs, credentials stored encrypted)
        sync_schedule: { type: string, description: "Cron expression" }
        is_active: { type: boolean, default: true }